refactor(api/story): change ID type from `Multi[int]` to `Multi[int64]` in `GetStoryCategoriesCount`

+++ api_story.go
@@ -954,13 +954,13 @@ type StoryCategory struct {
 // -----------------------------------------------------------------------------
 
 type GetStoryCategoriesCountRequest struct {
-	WorkspaceID *int        `url:"workspace_id,omitempty"` // [必须]项目ID
-	ID          *Multi[int] `url:"id,omitempty"`           // ID 支持多ID查询，多个ID用逗号分隔
-	Name        *string     `url:"name,omitempty"`         // 需求分类名称	支持模糊匹配
-	Description *string     `url:"description,omitempty"`  // 需求分类描述
-	ParentID    *int        `url:"parent_id,omitempty"`    // 父分类ID
-	Created     *string     `url:"created,omitempty"`      // 创建时间	支持时间查询
-	Modified    *string     `url:"modified,omitempty"`     // 最后修改时间	支持时间查询
+	WorkspaceID *int          `url:"workspace_id,omitempty"` // [必须]项目ID
+	ID          *Multi[int64] `url:"id,omitempty"`           // ID 支持多ID查询，多个ID用逗号分隔
+	Name        *string       `url:"name,omitempty"`         // 需求分类名称	支持模糊匹配
+	Description *string       `url:"description,omitempty"`  // 需求分类描述
+	ParentID    *int          `url:"parent_id,omitempty"`    // 父分类ID
+	Created     *string       `url:"created,omitempty"`      // 创建时间	支持时间查询
+	Modified    *string       `url:"modified,omitempty"`     // 最后修改时间	支持时间查询
 }
 
 // GetStoryCategoriesCount 获取需求分类数量
